fix(polymorphic): Split libs into node and browser env + update webpack config
