+++ src/lib/api/upload/file_tools.ts
@@ -14,11 +14,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { File as FsFile } from './file';
-import { isNode, SanitizeOptions, getMimetype } from './../../utils';
-import { FilestackError } from './../../../filestack_error';
-import fileType from 'file-type';
-
 export type RawFile = Blob | Buffer | File | string;
 export type NamedInputFile = {
   name?: string;
@@ -27,266 +22,19 @@ export type NamedInputFile = {
 
 export type InputFile = RawFile | NamedInputFile;
 
-const base64Regexp = /data:([a-zA-Z]*\/[a-zA-Z]*);base64,([^\"]*)/i;
-
-/**
- * Check if file is buffer
- *
- * @node
- * @param input
- */
-const isFileBuffer = (input: InputFile): input is Buffer => Buffer.isBuffer(input);
+export const base64Regexp = /data:([a-zA-Z]*\/[a-zA-Z]*);base64,([^\"]*)/i;
 
 /**
  * Check if file is blob
  * @param input
  */
-const isFileBlob = (input: InputFile): input is Blob => input.toString() === '[object Blob]';
-
-/**
- * Check if input is instance of browser file
- *
- * @browser
- * @param input
- */
-const isFileBrowser = (input: InputFile): input is File => input instanceof File;
-
-/**
- * Check if file is base64 string
- *
- * @param input
- */
-const isFileBase = (input: InputFile): input is string => {
-  if (typeof input !== 'string') {
-    return false;
-  }
-
-  if (input.indexOf('base64') > -1) {
-    input = input.match(base64Regexp).pop();
-  }
-
-  if (isNode()) {
-    if (Buffer.from(input, 'base64').toString('base64') === input) {
-      return true;
-    }
-
-    return false;
-  }
-
-  try {
-    return btoa(atob(input)) === input;
-  } catch (err) {
-    /* istanbul ignore next */
-    return false;
-  }
-};
+export const isFileBlob = (input: InputFile): input is Blob => input.toString() === '[object Blob]';
 
 /**
  * Check if file is instance of named interface
  *
  * @param input
  */
-const isFileNamed = (input: InputFile): input is NamedInputFile => input && input['file'] && input['name'];
+export const isFileNamed = (input: InputFile): input is NamedInputFile => input && input['file'] && input['name'];
 
-/**
- * Check if input is a valid file path
- *
- * @node
- * @param input
- */
-const isFilePath = (input: InputFile): input is string => require('fs').existsSync(input);
-
-/**
- * Convert encoded base64 string or dataURI to blob
- *
- * @browser
- * @param b64data     String to decode
- * @param sliceSize   Byte quantity to split data into
- * @private
- * @returns {Blob}
- */
-const b64toBlob = (b64Data: string, sliceSize = 512): Blob => {
-  let contentType = '';
-
-  if (b64Data.indexOf('base64') > -1) {
-    const matches = b64Data.match(base64Regexp);
-    b64Data = matches.pop();
-    contentType = matches[1];
-  }
-
-  const byteCharacters = atob(b64Data);
-  const byteArrays: any[] = [];
-
-  for (let offset = 0; offset < byteCharacters.length; offset += sliceSize) {
-    const slice = byteCharacters.slice(offset, offset + sliceSize);
-    const byteNumbers = new Array(slice.length);
-    for (let i = 0; i < slice.length; i += 1) {
-      byteNumbers[i] = slice.charCodeAt(i);
-    }
-
-    byteArrays.push(new Uint8Array(byteNumbers));
-  }
-
-  return new Blob(byteArrays, { type: contentType });
-};
-
-/**
- * Read file as array buffer
- *
- * @browser
- * @private
- * @param blob
- * @returns {Boolean}
- */
-const readFile = async (file): Promise<any> => {
-  /* istanbul ignore next */
-  if (!File || !FileReader || !Blob) {
-    return Promise.reject(new FilestackError('The File APIs are not fully supported by your browser'));
-  }
-
-  return Promise.resolve({
-    slice: function(start: number, end: number) {
-      return readPart(start, end, file);
-    },
-    release: () => {
-      file = null;
-    },
-  });
-};
-
-/**
- * Read file par instead of whole file to avoid browser crashing
- *
- * @param start - star byte
- * @param end  - end byte
- * @param file - file to slice
- */
-const readPart = (start: number, end: number, file): Promise<any> => {
-
-  return new Promise((resolve, reject) => {
-    const r = new FileReader();
-
-    const blob = file.slice(start, end);
-    r.onload = () => resolve(r.result);
-    r.onerror = reject;
-    r.readAsArrayBuffer(blob);
-  });
-};
-
-// =================== BROWSER UTILS ===================
-/**
- * Accepts b64string or blob file
- *
- * @browser
- * @param {*} fileOrString
- * @returns {Promise<File>}
- */
-const getFileBrowser = (input: InputFile, sanitizeOptions?: SanitizeOptions): Promise<FsFile> => {
-  let filename;
-  let file: Blob;
-
-  if (isFileNamed(input)) {
-    filename = input.name;
-    input = input.file;
-  }
-
-  if (isFileBrowser(input)) {
-    file = input;
-    filename = input.name;
-  } else if (isFileBase(input)) {
-    file = b64toBlob(input);
-  } else if (isFileBlob(input)) {
-    file = input;
-  } else {
-    return Promise.reject(new FilestackError('Unsupported input file type'));
-  }
-
-  return readFile(file).then(
-    async res => {
-      let mime = file.type;
-      if (!file.type) {
-        mime = getMimetype(await res.slice(0, fileType.minimumBytes), filename);
-      }
-
-      return new FsFile(
-        {
-          name: filename,
-          size: file.size,
-          type: mime,
-          slice: res.slice,
-          release: res.release,
-        },
-        sanitizeOptions
-      );
-    }
-  );
-};
-
-// =================== NODE UTILS ===================
-/**
- * Accepts Buffer or filepath or base64 string
- *
- * @node
- * @param {*} inputFile
- * @returns {Promise<File>}
- */
-const getFileNode = (input: InputFile, sanitizeOptions?: SanitizeOptions): Promise<FsFile> => {
-  // we have to wrap whole method inside promise because otherwise webpack is trying to parse
-  // all require statements inside for browser use
-  return new Promise((resolve, reject) => {
-    let filename;
-
-    if (isFileNamed(input)) {
-      filename = input.name;
-      input = input.file;
-    }
-    if (isFilePath(input)) {
-      let path = input;
-
-      return require('fs').readFile(path, (err, buffer) => {
-        if (err) {
-          return reject(err);
-        }
-
-        if (!filename) {
-          filename = require && require('path').basename(path);
-        }
-
-        return resolve(
-          new FsFile(
-            {
-              name: filename,
-              size: buffer.byteLength,
-              type: getMimetype(buffer, filename),
-              slice: (start, end) => Promise.resolve(buffer.slice(start, end)),
-            },
-            sanitizeOptions
-          )
-        );
-      });
-    }
-
-    if (isFileBase(input)) {
-      input = Buffer.from(input, 'base64');
-    }
-
-    if (isFileBuffer(input)) {
-      return resolve(
-        new FsFile(
-          {
-            name: filename,
-            size: input.byteLength,
-            type: getMimetype(input, filename),
-            // @ts-ignore
-            slice: (start, end) => Promise.resolve(input.slice(start, end)),
-          },
-          sanitizeOptions
-        )
-      );
-    }
-
-    return reject(new FilestackError('Unsupported input file type'));
-  });
-};
-
-export const getFile = (input: InputFile, sanitizeOptions?: SanitizeOptions) => (isNode() ? getFileNode(input, sanitizeOptions) : getFileBrowser(input, sanitizeOptions));
+export * from './file_tools.node';
