+++ webpack.config.js
@@ -24,6 +24,7 @@ const config =  {
     rules: [
       {
         test: /\.(js|jsx)$/,
+        exclude: /^.*\.node\.js$/,
         use: {
           loader: 'babel-loader',
           options: {
@@ -41,18 +42,20 @@ const config =  {
     ],
   },
   externals: [
-    // those externals are only used in nodejs
-    'fs',
-    'path',
+    'buffer',
     'crypto',
   ],
   plugins: [
-    // new webpack.NormalModuleReplacementPlugin(/^fs|path|crypto$/, './empty.js'),
     new CleanWebpackPlugin(),
     new webpack.BannerPlugin({ banner }),
     new webpack.DefinePlugin({
       'process.env.NODE_ENV': 'production',
       '@{VERSION}' : `${require("./package.json").version}`,
+    }),
+    new webpack.NormalModuleReplacementPlugin(/^.*\.node\.js$/,  (result) => {
+      if (result.resource) {
+        result.resource = result.resource.replace(/node/g, 'browser');
+      }
     }),
   ],
   devtool: 'source-map',
